added post location functionality

+++ FindFriends/Model/RequestModel/PostLocationRequest.swift
@@ -22,6 +22,6 @@ struct StudentDetailsRequest: Codable{
     let lastName   : String?
     let mapString  : String?
     let mediaURL   : String?
-    let latitude   : Float?
-    let longitude  : Float?
+    let latitude   : Double?
+    let longitude  : Double?
 }
